Changed Critical Links Row
